+++ lib/crypto/src/lib.rs
@@ -22,7 +22,7 @@ Common cryptographic procedures for a blockchain environment.
 extern crate alloc;
 extern crate core;
 
-pub mod biginteger;
+pub mod bigint;
 pub mod bits;
 #[macro_use]
 pub mod field;
